requirements.txt
# ── Core ───────────────────────────
Flask>=3.0
python-dotenv>=1.0
openai>=1.23
# ── DBpedia / SPARQL ───────────────
SPARQLWrapper>=2.0
rapidfuzz>=3.6
spacy>= 3.7
# ── Gunicorn (optional, prod) ──────
gunicorn>=21.2

# ── Caching ────────────────────────
redis>=5.0

